Restoring session spinner (#6)

+++ src/cli.rs
@@ -4,6 +4,6 @@ use clap::Parser;
 #[command(author, about)]
 pub struct Cli {
     /// Frame rate, i.e. number of frames per second
-    #[arg(short, long, value_name = "FLOAT", default_value_t = 60.0)]
-    pub frame_rate: f64,
+    #[arg(short, long, default_value_t = 60)]
+    pub frame_rate: u64,
 }
